requirements.txt
fastapi~=0.116.1
uvicorn[standard]~=0.35.0
transformers[torch]~=4.53.2
Pillow~=11.3.0
aiohttp~=3.12.14
cachetools~=6.1.0
pydantic-settings~=2.10.1
pydantic~=2.11.7